check for P by trying to run it
The current behavior of the function `checkPInstalled` hardcodes the
path to look for P. However it is possible that P is installed and
is not in this specific path. For instance:
- Users may have changed the
  [DOTNET_CLI_HOME](https://learn.microsoft.com/en-us/dotnet/core/tools/dotnet-environment-variables#dotnet_cli_home)
  environment variable
- Users may have P installed via [nix](https://nixos.org) or some
  other development environment manager

This has already caused issues in the past with issues like #75

Instead of trying to find where P is installed this pr simply tries to
run P. If the command `p --version` manages to run without throwing an
error it is assumed that P is installed on the system.