Add config option `p-vscode.compileOnSave`
Some users (such as myself) routinely save incomplete code. Running
a new task on each save event creates a lot of noise in such cases.
Even if there is an objection to this way of using a text editor, the
reality of muscle memory cannot be denied.

Whether this default behavior should be disabled is out of scope for
this change and merits an open discussion involving maintainers and users.